Representing Code / breakfast grammar
In the Representing Code chapter, I liked the
way that [breakfast, protein, cooked, bread]
could all be read as english before or after
substituting the generator. Meaning that
"cooked" is an adjective, and so is "scrambled",
and the generation is just an example of the variable.
I noticed that the crispiness didn't have that
same property, so I just switched it to "extra"
so it flows the same.

Also, I thought it would be nice to ensure that
the spacing works on the generators, so I added
spaces.